Create Cells
Cells can now be placed or removed by clicking the board.

Added a 2D array to keep the board state and a Cell component to keep
the positions of the cells.

Use an onClick event handler along with a helper function to retrive and
convert the click coordinate to a coordinate relative to the board.